registration and multy user functionality
* add registration

* add login page

* add some logic haha

* add list constrains

+++ src/Repository/CollectionRepository.php
@@ -3,8 +3,17 @@
 namespace App\Repository;
 
 use App\Entity\Collection;
+use App\Entity\User;
+use App\Exception\AccessException;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\Exception\RepositoryException;
 use Doctrine\Persistence\ManagerRegistry;
+use EasyCorp\Bundle\EasyAdminBundle\Config\KeyValueStore;
+use EasyCorp\Bundle\EasyAdminBundle\Context\AdminContext;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Core\Security;
+use Symfony\Component\Security\Core\User\UserInterface;
 
 /**
  * @extends ServiceEntityRepository<Collection>
@@ -16,13 +25,20 @@ use Doctrine\Persistence\ManagerRegistry;
  */
 class CollectionRepository extends ServiceEntityRepository
 {
-    public function __construct(ManagerRegistry $registry)
-    {
+    private ?User $user;
+
+    public function __construct(
+        ManagerRegistry $registry,
+        private readonly Security $security
+    ) {
+        $this->user = $security->getUser();
         parent::__construct($registry, Collection::class);
     }
 
     public function add(Collection $entity, bool $flush = false): void
     {
+        $entity->setOwner($this->user);
+
         $this->getEntityManager()->persist($entity);
 
         if ($flush) {
@@ -30,15 +46,21 @@ class CollectionRepository extends ServiceEntityRepository
         }
     }
 
+    /**
+     * @throws \App\Exception\AccessException
+     */
     public function remove(Collection $entity, bool $flush = false): void
     {
+        if ($this->user->getId() !== $entity->getOwner()?->getId()) {
+            throw new AccessException('Этот пользователь не может удалить эту коллекцию');
+        }
+
         $this->getEntityManager()->remove($entity);
 
         if ($flush) {
             $this->getEntityManager()->flush();
         }
     }
-
 //    /**
 //     * @return Collection[] Returns an array of Collection objects
 //     */
